--- examples/distributor/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "demo-distributor"
-version = "0.1.0"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-gstd.workspace = true
-parity-scale-codec.workspace = true
-
-[build-dependencies]
-gear-wasm-builder.workspace = true
-
-[dev-dependencies]
-gstd = { workspace = true, features = ["debug"] }
-gtest.workspace = true
-
-[features]
-debug = ["gstd/debug"]
-std = ["parity-scale-codec/std"]
-default = ["std"]
