--- examples/ping/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "demo-ping"
-version = "0.1.0"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-gstd.workspace = true
-
-[build-dependencies]
-gear-wasm-builder.workspace = true
-
-[features]
-debug = ["gstd/debug"]
-default = ["std"]
-std = []
