--- examples/vec/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "demo-vec"
-version = "0.1.0"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-gstd = { workspace = true, features = ["debug"] }
-
-[build-dependencies]
-gear-wasm-builder.workspace = true
-
-[features]
-default = ["std"]
-std = []
