+++ gclient/Cargo.toml
@@ -36,7 +36,6 @@ demo-calc-hash.workspace = true
 demo-calc-hash-in-one-block.workspace = true
 demo-custom.workspace = true
 demo-constructor = { workspace = true, features = ["std"] }
-demo-distributor.workspace = true
 demo-meta-io.workspace = true
 demo-new-meta.workspace = true
 demo-mul-by-const.workspace = true
