+++ gtest/Cargo.toml
@@ -26,7 +26,5 @@ log.workspace = true
 
 [dev-dependencies]
 demo-custom.workspace = true
-demo-piggy-bank.workspace = true
-demo-ping.workspace = true
 demo-futures-unordered.workspace = true
 demo-meta-io.workspace = true
