+++ pallets/gear/Cargo.toml
@@ -76,7 +76,6 @@ demo-async-init.workspace = true
 demo-async-recursion.workspace = true
 demo-async-tester.workspace = true
 demo-delayed-sender.workspace = true
-demo-distributor.workspace = true
 demo-init-fail-sender.workspace = true
 demo-init-wait.workspace = true
 demo-init-wait-reply-exit.workspace = true
@@ -110,7 +109,6 @@ demo-state-rollback.workspace = true
 demo-async-signal-entry.workspace = true
 demo-async-custom-entry.workspace = true
 demo-out-of-memory.workspace = true
-demo-ping.workspace = true
 demo-sync-duplicate.workspace = true
 demo-custom.workspace = true
 test-syscalls.workspace = true
